+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 6.3
 - Create Student Account: Fix reload page on School change, no AJAX in General_Info.inc.php
 - SQL add CREATE_STUDENT_ACCOUNT_DEFAULT_SCHOOL to CONFIG table in rosariosis.sql & Update.fnc.php
 - Add Create Student Account Default School option in Configuration.php & Help_en.php
+- Translate "Default School" & help to French & Spanish in rosariosis.po & help.po
 
 Changes in 6.2.3
 ----------------